Feat: 현재구독정보 컴포넌트 분리
