+++ src/components/myPage/CurrentSubscriptionInfo.jsx
@@ -1,0 +1,44 @@
+import styled from '@emotion/styled';
+import { Title, Text, Accordion, Box, Container } from '@mantine/core';
+import { SubscribeBadges, SubscriptionServices } from './index';
+
+const StyledContainer = styled(Container)`
+  background-color: ${({ theme }) => (theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[1])};
+  border-radius: 0.5rem;
+  padding: 1rem;
+`;
+
+const PresentSubscriptionFee = styled(Accordion)`
+  /* border-bottom: 1px solid white; */
+
+  button {
+    padding: 0;
+  }
+`;
+
+const CurrentSubscriptionInfo = () => {
+  console.log();
+  return (
+    <StyledContainer>
+      <PresentSubscriptionFee styles={{ item: { borderBottom: 'none' }, label: { padding: '0' } }}>
+        <Accordion.Item value="₩29,800">
+          <Accordion.Control>
+            <Title order={4}>현재 나의 구독료</Title>
+            <Text size="2rem">₩29,800</Text>
+          </Accordion.Control>
+          <Accordion.Panel>
+            <SubscriptionServices />
+          </Accordion.Panel>
+        </Accordion.Item>
+      </PresentSubscriptionFee>
+      <Box mt={16}>
+        <Title order={5} mb={10}>
+          구독하고 있지만 보고 있지 않아요
+        </Title>
+        <SubscribeBadges />
+      </Box>
+    </StyledContainer>
+  );
+};
+
+export default CurrentSubscriptionInfo;
