+++ src/components/myPage/SubscriptionServices.jsx
@@ -1,7 +1,7 @@
 import { Container, Title } from '@mantine/core';
 import SubscribeBadges from './SubscribeBadges';
 
-const SubscriptionInfo = () => (
+const SubscriptionServices = () => (
   <Container mt={10} p={0}>
     <Title order={5} mb={10} fw={400}>
       현재 구독하고 있어요
@@ -10,4 +10,4 @@ const SubscriptionInfo = () => (
   </Container>
 );
 
-export default SubscriptionInfo;
+export default SubscriptionServices;
