+++ src/components/myPage/index.js
@@ -4,4 +4,5 @@ export { default as Collection } from './Collection';
 export { default as Collections } from './Collections';
 export { default as SubscribeBadges } from './SubscribeBadges';
 export { default as Statistics } from './Statistics';
-export { default as SubscriptionInfo } from './SubscriptionInfo';
+export { default as SubscriptionServices } from './SubscriptionServices';
+export { default as CurrentSubscriptionInfo } from './CurrentSubscriptionInfo';
